requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
matplotlib>=3.7.0
